Excluído a classe ProviderRestBase e alteração do método Enviar no ProviderBase para nao permitir override
